Add files via upload
The first lecture slides, which we could not display due to a projector issue (first they didn't turn it on, then it wasn't set correctly to take any input apparently, will resolve with IT before second lecture). I could have pushed the change and then everybody follow on their laptops but it didn't occur to me at the moment. Apologies these could not be up before lecture.